django/requirements.txt
django
django-redis
redis
django-storages
boto3
pillow
django-cors-headers
